requirements.txt
haystack-ai
ollama-haystack
chromadb
chroma-haystack
